+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## UNRELEASED
+- Closes dialog overlay only when needed
+- Auto prefixes style passed to `getComputedStyle()`
 - Exports all classes using `export default` ES6 syntax
 - Renames all component CSS class name with the prefix `cc-`
 - Renames class `debug` to `debugging`